+++ fishTerm/fishTerminal.sh
@@ -2,17 +2,25 @@
 sudo apt update
 sudo apt upgrade
 if [ -d "$HOME/.config/omf" ]; then
-  echo "omf is already installed"
-  exit
+	echo "Everything installed!"
+	exit
 fi
-sudo apt install fish
+if [ ! -d "$HOME/.config/fish" ]; then
+  sudo apt install fish
+else
+  echo "fish is already installed!"
+fi
+if [ ! -d "$HOME/.config/omf" ]; then
 read -r -d '' myscript <<'EOF'
 import sys
 import os
 for line in sys.stdin:
-    if "successfully installed" in line:
-        os.system("killall fish")
+	if "successfully installed" in line:
+		os.system("killall fish")
 EOF
-
-fish -c "$(curl -L https://get.oh-my.fish)" | python3  -c "$myscript"
-echo "omf install bobthefish" | fish
+	fish -c "$(curl -L https://get.oh-my.fish)" | python3  -c "$myscript"
+	echo "omf install bobthefish" | fish
+else
+	echo "omf is already installed!"
+fi
+exit
